# rest-server/doc/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "DM124 REST API",
    "description": "Trabalho Prático da disciplina DM124 - Lucas Leon Labanca"
  },
  "host": "localhost:8080",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic",
      "in": "header",
      "name": "Authorization",
      "description": "Basic Auth Base 64"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/users/signup": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Add a user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "description": "User",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserToAdd"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User added succesfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Get a list of all of the users",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/UserList"
            },
            "description": "User list found"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request: error or not allowed to users with role: USER"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Get a user by ID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/User"
            },
            "description": "User found"
          },
          "400": {
            "description": "Bad Request: error or user with role USER is not authorized to get someone elses user"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "description": "Update a user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "User ID"
          },
          {
            "name": "User",
            "in": "body",
            "description": "User",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserToAdd"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/User"
            },
            "description": "User updated succesfully"
          },
          "400": {
            "description": "Bad Request: error or user with role USER is not authorized to update someone elses user"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "Delete a user by ID",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          },
          "400": {
            "description": "Bad Request: error or user with role USER is not authorized to delete someone elses user"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "admin01"
        },
        "role": {
          "type": "string",
          "example": "ADMIN"
        },
        "updatedAt": {
          "type": "string",
          "example": "2021-09-24T00:45:54.199Z"
        },
        "createdAt": {
          "type": "string",
          "example": "2021-09-24T00:45:54.199Z"
        }
      }
    },
    "UserList": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 2
          },
          "name": {
            "type": "string",
            "example": "user01"
          },
          "role": {
            "type": "string",
            "example": "USER"
          },
          "createdAt": {
            "type": "string",
            "example": "2021-09-24T00:48:14.123Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2021-09-24T00:48:14.123Z"
          }
        }
      }
    },
    "UserToAdd": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "admin01"
        },
        "password": {
          "type": "string",
          "example": "123456"
        },
        "role": {
          "type": "string",
          "example": "ADMIN"
        }
      }
    }
  }
}